Update Distance_Implementation_Sale_Water.R
The parameter "Great Circle" in the function st_distance (part of the "sf" library) allows recovering distances (corrected for earth sphericity). The output is a list of matrices (m x n) - with the distances calculated between all the locations in the second object and one of the locations in the first object (one at a time)
- The corresponding nearest distance can then be simply recovered by running: apply (output from st_distance, min) or another type of iteration.
For the two sf objects (sales and water use), the code crashed when calling st_distance(…), reporting an output matrix > 50 GB

A simple apply-iteration ran with the "Great Circle" or haversine-distance, coded as a function apart, performed all the distance calculations in less than 5 seconds.